Remove dichotomy Tree/TreeNode by removing TreeNode.
We had two classes that were the same: Tree and TreeNode.

The way we represent a tree, a Tree can be either a single node or
a node conected to others (children and/or parent). So there is no
distinction between node and tree.

Having two classes doing the same is an unnecessary complication that
can bring difficulties (grepping for example, or mixing it with the
TreeNode from skbio), with no clear advantage if at all. So I am
removing TreeNode in favor of Tree.

There are some additional cosmetic changes.