tree.pyx: Nicer printing of tree.
The __str__() method now uses a nicer text art to show the tree.

It can use arbitrary x and y padding.

It cannot show the internal nodes for the moment, but I am leaving
this out until I find if we want to substitute the default printing
for this one. Also, it is not the common case.