Last commit message was wrong - It should be: Color the label of the search button with the color of the search results.
It referred to another change I am working on.

The previous commit was about coloring the label of the search button to match the results.

This commit is just to comment on it. The idea with the colored button is that
we can have many searches and see which is which easily.

Kudos for the idea to Claudia.