faces.py: Comment on the hacks made for LayoutBarPlot to work and minor fixes.
The width argument is the hack. The height argument is not used at all, so I
remove it. The rotation argument works but makes the boinding box be incorrect,
so we will have to fix this in the future.

The tooltip is already stored when initializing CircleFace, so we should not
add it again.