Remove _properties and use props directly.
The only advantage to have an internal _properties and then a
setter and getter for props was that we could raise an error
if something different from a dict was assigned to props. But
this is a small gain, and is actually unnecessary! Since tree.pyx
is a cython module, it already checks and raises an exception if
we try to assign a non-dict to props.

This change simplifies a bit the code - no worries of having a
distinction between props and _properties.