server.py: Fix (remove) calls to unexistent get_popup_props()
There was a Tree method called get_popup_props(), which seems to
have been removed in commit 464710ebaab. But after removing it,
it was still called from a couple of places in server.py.

It seems it was just to return some information, that is not
actively being used by anything else for the moment.

This commit fixes the calls by removing them. They would produce
an internal server error when selecting a clade.

By the way, it is unclear to me what the /find endpoint is for.
There seem to be no use or tests for it anywhere.