Remove example evol data from the repository (moved to ete-data).
We had quite a few data files, and even duplicated, which are better
in a separate repository. So we have moved them to another repository:
https://github.com/etetoolkit/ete-data/tree/main/examples/evol_data

Also, there are some comments in examples/evol/README, and some updates in the
test at ete4/test/test_ete_evol.py (but the test does not work, before or after
the changes -- it just is closer to being fixed).