Add update_ete_data() and use it to download/update data files when needed.
update_ete_data() will refresh the given path with the given url. The path,
if relative, is resolved to ETE_DATA_HOME.

Also, now that we are moving the big data files outside of the repository,
the two layout files in ete4/smartview/renderer/layouts that were loading
big jsons (domain_layouts.py and ncbi_taxonomy_layouts.py) will use
update_ete_data() to get them from the ete-data repository.

Note that in the future we will also move out all the layout files
except the ones that are basic and or very general.