gui.js: Fix setting the name of the tree in view.tree.
`view.tree` is supposed to store the name of the tree (a string),
and `trees[view.tree]` should contain its id (an int). See for
example:

const trees = {};  // will translate names to ids (trees[tree_name] = tree_id)

In commit ad70ed751 it changed from
  view.tree = value;
to
 view.tree = trees[value] || value;
but there does not seem to be a good reason for that. I can not find
any explanation in the logs, and it seems like a bug.

This commit fixes changing to another tree in the url with ?tree=NAME, like in:
  http://localhost:5000/static/gui.html?tree=example
when that tree has been added with a POST, for example, as with
  http POST http://localhost:5000/trees name=example newick='(a,b);' id=444