server.py: Fix the properties shown in a node when visualizing a tree from the console.
In commit 2d1d63a17a6dae48fded6a3a20a28404427ec4d8 we fixed that
popup_prop_keys could be passed when adding a tree with a POST,
but that change introduced a bug: if opening a tree from the console
with tree.explore() without popup_prop_keys, which is normal, it
would now be empty.

This commit fixes it and makes the tree have
popup_prop_keys=DEFAULT_POPUP_PROP_KEYS if they are not passed.