Add file_extract() to guess what the given string is.
We will want to use it in places like the constructor of Tree,
or when creating trees from "newicks" and so on.

The function uses simple heuristics to see if a string is a
the name of a file that contains the data, or it is the data
itself.

I do not like the approach (I prefer to have the user pass
explicitely either a file name or the newick text, or whatever
text). But if we are going to use it, I think it is better
in this way than within each of the parsers.