Give a clear message when accessing an endpoint that does not exist for the requested method.
For example, if we accessed http://127.0.0.1:5000/layouts/update with a browser,
a GET request was made and that raised an uncaught exception.

This change should make more clear from the perspective of the user what is
going on. That is, they accessed an endpoint that does not exist for the
method (GET / PUT / ...) used.