Update to pyqt6 and many related changes -- see long description.
The main change in this commit is an update from pyqt4/5 to pyqt6.

Also, treeview (the visualization based on qt) is now an optional
dependency. That means that by default the full qt will not be
downloaded with ete, unless you want to use treeview.

We moved the content of svg_colors.py to utils.py.

We removed several deprecated keywords like "line_type".

Finally, there are some simplifications and cosmetic changes all around.

The changes are made in part to address the following PRs:

- https://github.com/etetoolkit/ete/pull/700
- https://github.com/etetoolkit/ete/pull/699
- https://github.com/etetoolkit/ete/pull/698