Remove sdoc/ directory in preparation for an up-to-date documentation.
The contents of sdoc/ have a lot of deprecated and non-working stuff.
We are putting the new documentation together at the doc/ directory now,
and the contents of sdoc/ can only make it confusing.

So, while we keep a historical copy somewhere to see what to port to ete4,
we are removing the sdoc/ directory.