Several api changes like  attr -> prop and more (see description).
The changes in the api:

- treediff(attr1, attr2, ...) -> treediff(prop1, prop2, ...)
- sort_descendants(attr) -> sort_descendants(prop)
- get_cached_content(store_attr, ...) -> get_cached_content(prop, ...)

And there is a new implementation of t.get_cached_content() that is
hopefully easier (and slightly faster, and probably fixes a bug
when container_type=list and there was a single name instead of a list
of attributes/properties given.

Also, get_cached_content() now only assigns the node themselves
or a single attribute, which is the most common case. If you want
to get several attributes (like before with store_attr=[p1, p2, ...]),
simply get the nodes themselves (with prop=None, the default), and
get the properties of those nodes when needed.