phylotree.py: Several fixes, mainly when sp_naming_function=None.
When sp_naming_function=None it was not really setting it to None
in all the children.

Actually, the way that the constructor was calling the constructor
of Tree was bad. And several things were set in an irrelevant (and
confusing) way.

Now we call the constructor of Tree() with the proper arguments and no
more. And we have fixed the set_species_naming_function(), which did
not remove it if it was None.

We also raise an exception if we try to set n.species for a node that
has a species naming function.

And in general, a cleanup of the code.