Remove unnecessary reuse_names=False in calls to t.populate().
reuse_names=False is already the default in t.populate().

Also, nobody is going to want to use it as True, because
it gains basically nothing, and it is not an scenario where
you would want to use populate().

So that argument is more confusing than helpful. And this
change (which really changes nothing) is in preparation
for a removal of that argument from the function itself
that I intend for the next commit.