tree.pyx: Improvements to t.populate(), specially the use of short names.
Instead of the itertools.combinations_with_replacement(charset, 10)
which produced quite long and ugly names, this version produces the
shortest possible names with the a-z chars. I think they look better.

Also, this change removes the reuse_names argument, that is polluting
the api more than being helpful.

And there is some cleanup of the code and documentation.