Remove inheritance diagram from the doc.
We do not have a lot of classes inheriting from others.
In fact, we have about 3.

So the diagram does not help much, and to show them you
have to have a correct installation of graphviz or
something like that, which is not currently working at
https://etetoolkit.github.io/ete/ (where we have the
documentation automatically updated), and it is not worth
changing that.

Otherwise, https://etetoolkit.github.io/ete/reference/index.html
started with:
  digraph inheritance9e90e02d16 { bgcolor=tr...
and so on. Terrible. For nothing.