pyproject.toml: Add missing dependency for requests.
I used it in ncbiquery.py, but did not add it as a dependency.

The requests library (https://github.com/psf/requests) is cool and
I think we should use it as much as possible for all our http
connections. And is well maintained, widely available and small.

So here it comes.