tree.pyx: t.explore() has include_props with name and dist by default.
So, if we open a big tree with a lot of annotations, ete will not show
them with just t.explore(). Instead, you need to explicitely do
t.explore(include_props=None) to see them all.

The reason is that big annotated trees would go slowly otherwise, if
opened with the defaults. And we want the default to be fast, so even
if a casual user with a small tree will not see the annotations, that
is on them, rather than on a more advanced user with a huge tree.