coretype -> core
The part in ete4/coretype is not only about types. It is about
core functionality. So this name reflects it better.

And coretype always seemed off to me. This hopefully makes the
content of the directory (submodule) more clear: the basic things
related to Trees, over which everything else is constructed.

We will use it soon to move functionality out of tree.pyx and
gardening.pyx and put it in tree_operations.pyx or something
like that.