tree.pyx: Cleanup of the traverse functions.
They do not traverse the descendants only, but all the tree including
the root. The names of internal functions and docstrings have been
changed to reflect this.

Also, the order in which they appear is more consistent, and a
reference to the standard tree searches (depth-first and breadth-first)
has been added.