setup.py: Automatically get all the pyx files for cythonize.
This way, it still works with
  pip install -e .
and with
  python3 setup.py develop
as it used to, but in addition to that, it works with
  python3 -m build
which did not work before.

We need to run  python3 -m build  (which creates a couple of files in
the directory dist/), as a step to upload the package to pip.
See https://packaging.python.org/en/latest/tutorials/packaging-projects/