Take ultrametric out from TreeStyle.
Modifying a tree should not be part of a "style". It was also causing
really tricky problems with the initialization of a tree as ultrametric
and then properly modifying it at any point in the explorer.

This commit removes the attribute "ultrametric" from TreeStyle
and changes the confusing "tree" name for both the Tree structure
and the data associated to it, that is used all around in server.py.

The total information for a tree is called tree_data now, and the
tree (of class Tree) itself is tree_data.tree. Also, having forced
it to be ultrametric is stored now in tree_data.ultrametric, and not
on any style.

There are some minor fixes and simplifications too.