On t.explore(), try to use an available port by default.
Instead of the ete graphics server listening on 5000 by default,
we now try to find a port that is available and use it.

This would help with a lot of cases where people are having problems
with ete and they launch several interpreters, or do not kill old
ones, and then they run into occupied ports all the time.