Fixes in t.resolve_polytomy().
The description of the function was inaccurate (the topology is
arbitrary, not random).

It also got parameters that should not be there ("defaults", and
instead it now sets all new nodes to dist=0 and support=0 as they
should, and makes sure that support=0 for all the polytomy children).

The order of the created ladder was a bit ugly:

      c
      d
    b
  a

Now it is:

      a
      b
    c
  d

And a few more simplifications.