Add t.render_sm() to create an image with a tree without launching a browser.
This uses the optional module selenium, which behind the scene
calls chrome in headless mode and makes a snapshot.

To do it cleanly, we have to manage the server ourselfs, instead of
letting bottle do it (with bottle.run). This way, we can stop it
at any time (server.server_close() and server.shutdown()).

To be able to choose the verbosity, we have to modify
WSGIRequestHandler and override its log_request function. This is
quite annoying, and less clean than using
  run(quiet=not verbose, ...)
that we did before, but when managing the server ourselves, this
was the best way I could find to do it.