draw.py: Do not draw on branch top/bottom if there is no space.
We use the convention dx=0 to say that there are no constraints in
the x width, for example in positions "right", "aligned" and "left".
But for positions "top" and "bottom", dx=0 can happen if we have a
branch with length 0, and so this change will avoid drawing things
there because it confuses dx=0 with "as much space in x as you want".