Allow drawing circles of any radius.
We will want to draw cicles for more things than just the nodedots.
This change allows to draw them, while keeping the radius of node
dots controlled on the gui through the value in the global object
"view".

By the way, we may want to change the value of radius for node dots
programatically, and in that case we may want to change this code
so it can be done. A posible way is to have the radius be some
special value like -1, 0 or None when we want to control it through
the gui, and a poitive value otherwise.