draw.py: Allow layouts node_decos() to return None or single elements too.
Before, the function node_decos() of a layout had to return a list
of Decorations.

With this change, the function can return None, a single Decoration,
a single Face, or a list (iterable in general) of Decorations or Faces.

All the Faces will be wrapped automatically into Decorations if they
are not already.

This allows to write simpler node_decos() functions. For a very
simple example:

def node_decos(node):
    if node.is_leaf and node.dist > 1:
        return TextFace("I am a long leaf")