layout.py: Simplify default_node_decos().
This is about the function used to get the decorations of a node in
the default layout.

It used to return a list of decorations, but since all we care is
about iterating over it, we can convert it to an iterator and
have a cleaner syntax. So we do not construct a list anymore, we
just yield the Decorations whenever is appropriate.