draw.js: Use new functions tree2rect() and tree2circ().
They substitute two old functions:
  transform() -> tree2rect()
  cartesian_shifted() -> tree2circ()

But they have a better name (hopefully), are grouped together, and
are now used extensively in the rest of the code whenever such
transformations are needed.

The resulting code is a bit more readable, I think.

Also, we use tree2circ() in create_circle(), which together with the
previous change in the drawing of the nodedot, fixes the position
of nodedots in the graphics.