Draw aligned items in circular mode right after the furthests drawn non-aligned item.
We used to have a crude approximation where the aligned items in
circular mode would start at r=tree_size.width (that is, the distance
of the furthest leaf).

Now we changed it so they will start after the r of the furthest item
that was drawn before the aligned items.

The way we implement this is by introducing a new command, set_xmax().
The drawer in the backend keeps track of the maximum x (or r in
circular mode) that has been reached (in the self.xmax_reached
variable), and after drawing everything sends it with the set_xmax()
command. The javascript code then uses that value to translate the
items in the aligned panel.