Allow "parser" to be a string too, in addition to integers or dicts.
For the newick format, we call "parser" a dictionary that has
information about how to process the leaf nodes and the internal
nodes. But we used to accept old-style integers (parser=1, parser=2,
etc) as ete used to do.

Now we want to accept strings too. For example, the name of a program
that outputs the support value as "N/M" and so on.

This change will allow to create predefined parsers for them. It also
simplifies things a bit.