menu.js: Allow to modify the min_size_content from the gui too.
The min_size_content controls when some drawing element is too
small to appear (so it either does not appear or a box appears
in its place).

This is different from the min_size parameters, which controls
when a node is too small to appear by itself, so it is then
collapsed (possibly with other nodes).