Rename "styles" to "node_styles" in the tree style.
So it is more clear that the element "node_styles" refers to the
styles that can be referenced when drawing the nodes.

That is, in the tree style dict, one can add names of styles like:
  tree_style = {"node_styles": {"s1": {...}, "s2": {...}}}
and then, a node can draw elements using style="s1" for example, and
that will automatically set its style with all the values in
node_style "s1".