draw.py: Change the order of (dis)appearance of decorations in a column.
If we have a total dy vertical space in a column, and there are n rows
of decorations to place, the first decoration is granted a space of:

  dy / n

The second one, the remaining space after the first one has been
drawn, divided by the remaining decorations to draw. And so on.

That makes the last drawn decorations to have the maximum possible
available space, because the previous ones may have used less space
than what they had allocated. They may even not have been drawn at all
if the space was too small!

But the effect that this caused could be surprising, because the first
decorations may not appear (not enough space), while the last do
appear (enough space, because nothing was drawn before). We want the
contrary: if any decoration is not going to appear for lack of space,
it better be the last ones! This commit implements that change.