layout.py: Tree styles are added to the default tree style automatically.
We want to keep the node_styles defined in DEFAULT_TREE_STYLE even if
we are adding other node_styles and so on.

To force the user to do add_to_style() manually seems a bit too much.

This change makes the default behavior to work pretty nicely: new tree
styles are added to the default tree style.