Fix initial values of nodebox opacity and color.
They correspond now to the ones defined in the css (gui.css) for
nodebox.

This way, when we start changing values in the menu, the nodes color
and opacity only change for the ones that have not been set in
draw_node() (that is, for the ones that have default values from the
css or from the tree_style).