draw_node(node), draw_collapsed(nodes) -> draw_node(node, collapsed).
As suggested by Jaime, instead of having in the layouts these two functions:

  draw_node(node)
  draw_collapsed(nodes)

we now have a single function:

  draw_node(node, collapsed)

where the second argument is a list of sibling nodes collapsed together, if we
have things that are collapsed.

This lets the user write more comfortably all the logic for the
faces and styles that they want in the different cases.

Also, we let the user write a function with just:

  draw_node(node)

In that case, when we draw collapsed nodes we actually apply that function
to all the collapsed siblings and add all the representations from them.