Add node dots of many shapes, and styled in multiple places.
Now a node dot can be a circle, triangle, etc. And we can set
its shape and its radius as a style property, which works
consistently with others (if it is a tree_style property, it
can be changed from the gui, if it is a node_style property,
it is fixed, and so on).

This requires looking into the styles and extracting values,
with the new function pop_style() in draw.js, and the modified
set_tree_style() in gui.js.