Change internal naming outline -> collapsed in various places.
It used to be that we represented the collapsed nodes with
an outline (which was a triangle extending as far as the full
node).

Then, we used a skeleton of the subtree as a representation of
the collapsed nodes, but we kept the name "outline" for many
of the functions related to that.

This change clarifies it a bit more in the code. Also, it is in
preparation to a new style property for the collapsed nodes,
which will allow to choose between an outline representation
as it was before, and a skeleton one as it is now.