Save original styles and restore them before set_tree_style() overrides them.
This way we can go back to an empty/default style if we deactivate layouts.

It seems also the proper way to change things: revert them first to
the default values, and only then change the ones that have been set
in a tree style.

Also, there is a fix when checking if the shape of a node has a valid
value (in set_tree_style()):

-            if (! typeof shape === "number" &&
+            if (! (typeof shape === "number") &&