In set_tree_style(), update the global variables that track the style too.
Use global variables in the fields of the `view` object in gui.js with
names that correspond to the css/svg properties that we can change,
like "fill", "stroke", "stroke-width".

This, in addition to the update that we do in gui.js:set_tree_style(),
allows us to update the menus and the CSS directly.

Also, change the internal names:
  line.length, line.children -> hz_line, vt_line
so the names of the lines are more consistent with the rest of the code.