Allow to pass extra style arguments in t.explore().
So now we can do things like:
  t.explore(is_leaf_fn=lambda node: node.level > 3,
            include_props=None,
            exclude_props=["dist"])
and it will work the same as if we had passed them in the style
of the last layout.

It does this by creating a new layout if there are more keyword
arguments at the end of explore(). This new layout, named
"extra arguments" (which can be selected/deselected from the gui),
adds those arguments as a dictionary to its tree_style.

This allows the behavior that Jaime wanted, and the one I preferred,
and also allows to pass any tree style as argument without having
to create a Layout for it.