tree_style, node_style -> draw_tree, draw_node
draw_tree() takes a tree and returns a list of dictionaries and
decorations, the same as draw_node(). But when defining a layout, we
can assign it to a dictionary, and it will create a function that
returns it appropriately.

Also, when we do t.explore(...) we can pass as arguments any tree
style, and it will be incorporated in a special layout called "extra
arguments" that contains those options.

Finally, there is a fix for the radius of a dot in circular mode, so it
is as big as specified but not bigger than the node height.