Add active option for Layouts.
I am not satisfied with this, because this is probably not
the right way to put it.

Whether a layout is active or not should not be a property of
the layout itself, I think.

Anyway, this behavior is the same as it had before in ete4,
and whether the layout is active or not can be changed from the
gui. The argument only affects the initial state of the layout
when viewing a tree.

We may want to revisit all this in the future.