Fix drawing of negative x axis when in ciruclar mode with zoom.
When we zoom over the angle and items get hidden in circular mode
(because they are beyond an angle of [-180, 180]), we show a nice
horizontal line so the user knows what is happening.

After an old change to the names of css applied styles, the line
looked like a thin black line. This change makes it look cooler.