Add sequence type for nucleotides, and allow specifying render method.
When drawing a sequence now we can specify if it is of seqtype aa
(amino acids) or nt (nucleotides). The colors will change accordingly.

Note that the pixi version is missing the color change for
nucleotides. Only the svg version is complete at the moment.

Also, added an option "render" to the sequence drawing, which can be:
auto, raster, svg. And a selector in the advanced menu in the gui to
override how they are drawn too if we want.