Add rotations to TextFace and similar faces.
Now we have the argument rotation=... (an angle in degrees) for all
these faces:
- EvalTextFace
- PropFace
- TextFace

The text will be rotated, but it is not taken into account for the
rectangle that marks the size that it occupies. This may not be what
we want in the future, but for the moment it is a nice feature for the
future headers and footers that we will add.

Also, this commit includes in graphics.py the function draw_polygon()
and in draw.js how to translate polygons. This is in preparation for
the next commit, that will introduce PolygonFace.