Add indent parser.
This parser reads one of the most common informal formats that
are used to represent trees.

For example, the output of the commands:
  pstree -a
  tree
  eza --tree
and many more, including when ete prints a tree with:
  t.to_str(cascade=True)

It is fun to see those trees in the explorer!

Also, added some tests and simplified a bit some code.